requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
transformers==4.39.3
torch==2.2.1
openai==0.28
google-generativeai==0.8.4